Implement an API and update the SQL.
The SQL now returns more data from nearby_apartments.
The API exposes a GET request that returns 100 apartments max.
The API also exposes a POST request that returns a maximum of 100 apartments by default but can also be filtered using min/max for sq_footage, rooms, and price. It can also filter based on max distance in meters. But requires a lat and long coords to do so.