Improve the /rent page.
This update takes advantage of the new SQL that exposes the first/last name. It's now shown on the listing. It also includes some bug fixes with mixed up long/lats.

This update also uses a state to store the selection so it updates the nearby listing of its parent when needed.